wip In mocha-dark.js save scheme on toggle only
The color scheme setting is not saved in storage/cookie unless the user clicks on the circular progress indicator.
This allows the CSS prefers color scheme and body.className values to take effect.
Once the user clicks the progress toggle the setting is stored and remains.
Using storage, works for both file:// and http:// protocols so can probably discard cookie setting.